fix: Don't vertically parallelize cse contexts
