+++ crates/polars-mem-engine/src/executors/stack.rs
@@ -33,22 +33,8 @@ impl StackExec {
                         self.has_windows,
                         self.options.run_parallel,
                     )?;
-                    if !self.options.should_broadcast {
-                        debug_assert!(
-                            res.iter()
-                                .all(|column| column.name().starts_with("__POLARS_CSER_0x")),
-                            "non-broadcasting hstack should only be used for CSE columns"
-                        );
-                        // Safety: this case only appears as a result
-                        // of CSE optimization, and the usage there
-                        // produces new, unique column names. It is
-                        // immediately followed by a projection which
-                        // pulls out the possibly mismatching column
-                        // lengths.
-                        unsafe { df.get_columns_mut().extend(res) };
-                    } else {
-                        df._add_columns(res, schema)?;
-                    }
+                    // We don't have to do a broadcast check as cse is not allowed to hit this.
+                    df._add_columns(res, schema)?;
                     Ok(df)
                 });
