+++ py-polars/tests/unit/test_cse.py
@@ -765,3 +765,21 @@ def test_cse_non_scalar_length_mismatch_17732() -> None:
     )
 
     assert_frame_equal(expect, got)
+
+
+def test_cse_chunks_18124() -> None:
+    df = pl.DataFrame(
+        {
+            "ts_diff": [timedelta(seconds=60)] * 2,
+            "ts_diff_after": [timedelta(seconds=120)] * 2,
+        }
+    )
+    df = pl.concat([df, df], rechunk=False)
+    assert (
+        df.lazy()
+        .with_columns(
+            ts_diff_sign=pl.col("ts_diff") > pl.duration(seconds=0),
+            ts_diff_after_sign=pl.col("ts_diff_after") > pl.duration(seconds=0),
+        )
+        .filter(pl.col("ts_diff") > 1)
+    ).collect().shape == (4, 4)
